Shut up compiler warning.

+++ src/InputOutput.cpp
@@ -429,7 +429,7 @@ void IO::Write(const char *p, size_t size)
     if (Is_Compressed())
     {
 #ifdef COMPRESS_OUTPUT
-        const int error_code = gzwrite(compressed_fh, p, size);
+        const int error_code = gzwrite(compressed_fh, p, (unsigned int)size);
         assert(error_code != 0);
 #else
         std_cout << "Can't be here!!! (" << __FILE__ << " line " << __LINE__ << "). Aborting.\n" << std::flush;
@@ -473,7 +473,7 @@ void IO::WriteString(const std::string &format, ...)
         if (Is_Compressed())
         {
 #ifdef COMPRESS_OUTPUT
-        const int error_code = gzwrite(compressed_fh, string_to_save, strlen(string_to_save));
+        const int error_code = gzwrite(compressed_fh, string_to_save, (unsigned int)strlen(string_to_save));
         assert(error_code != 0);
 #else
         std_cout << "Can't be here!!! (" << __FILE__ << " line " << __LINE__ << "). Aborting.\n" << std::flush;
